fix(collapse-diff): when diff change is only a file rename
The diff's structure is different when the change is only
a file rename. Previous behavior didn't take that into
account and errored out in those cases.
